Add chapter 11
CHapter is the next part of chapter 10 as it showcases the basic use
maps.

General reminder: Collections in Kotlin are read-only (immutable)
by default. In order to change, one needs to declare or convert
the collect the collect to a mutable one.

Note : Chapter 9, was not a hands-on chapter. It demonstrates use of
the standard library functions such as: let, apply, run, with, also,
takeIf and takeUnless.