Add chapter 13
Chapter13 is about Object initialization, meaning, how to use constructors,
how to initialize properties (inline in a primary constructor, at declaration,
in secondary constructor or in an init block), and how to do late and lazy
initialization. One have be really careful on the instructions' order as it
can lead to unexpected behavior.