Add chapter 16
chapter 16 is about interfaces and abstract classes. Both pretty much
work the same as in JAVA. Interfaces are for declaring behavior (without
implementation) or states that concrete classe will implement. Abstract
classes are for having a main class, usually in a hierarchy, with abstract
or/and concrete behavior; they can also have states.